Crate revmc_builtins

revmc-builtins

EVM bytecode compiler builtins.

Modules

gas

Gas calculation utilities.

ir 🔒 ir

macros 🔒

utils 🔒

Structs

Builtinsir

Builtin cache.

Enums

Builtinir

Builtins that can be called by the compiled functions.

CallKind

The kind of a *CALL* instruction.

CreateKind

The kind of a CREATE* instruction.

ExtCallKind

The kind of a EXT*CALL instruction.

Constants

EXTCALL_LIGHT_FAILURE

The result of a EXT*CALL instruction if the gas limit is less than MIN_CALLEE_GAS.

Functions

__revmc_builtin_addmod⚠

__revmc_builtin_balance⚠

__revmc_builtin_blob_base_fee⚠

__revmc_builtin_blob_hash⚠

__revmc_builtin_blockhash⚠

__revmc_builtin_call⚠

__revmc_builtin_calldatacopy⚠

__revmc_builtin_codecopy⚠

__revmc_builtin_codesize⚠

__revmc_builtin_create⚠

__revmc_builtin_data_copy⚠

__revmc_builtin_data_load⚠

__revmc_builtin_difficulty⚠

__revmc_builtin_do_return⚠

__revmc_builtin_eof_create⚠

__revmc_builtin_exp⚠

__revmc_builtin_ext_call⚠

__revmc_builtin_extcodecopy⚠

__revmc_builtin_extcodehash⚠

__revmc_builtin_extcodesize⚠

__revmc_builtin_func_stack_grow⚠

__revmc_builtin_func_stack_pop⚠

__revmc_builtin_func_stack_push⚠

__revmc_builtin_gas_price⚠

__revmc_builtin_keccak256⚠

__revmc_builtin_log⚠

__revmc_builtin_mcopy⚠

__revmc_builtin_msize⚠

__revmc_builtin_mulmod⚠

__revmc_builtin_panic⚠

__revmc_builtin_resize_memory⚠

__revmc_builtin_return_contract⚠

__revmc_builtin_returndatacopy⚠

__revmc_builtin_returndataload⚠

__revmc_builtin_self_balance⚠

__revmc_builtin_selfdestruct⚠

__revmc_builtin_sload⚠

__revmc_builtin_sstore⚠

__revmc_builtin_tload⚠

__revmc_builtin_tstore⚠