revmc_builtins::gas

Function calculate_initial_tx_gas

pub fn calculate_initial_tx_gas(
    spec_id: SpecId,
    input: &[u8],
    is_create: bool,
    access_list: &[AccessListItem],
    authorization_list_num: u64,
) -> InitialAndFloorGas

Initial gas that is deducted for transaction to be included. Initial gas contains initial stipend gas, gas for access list and input data.

Returns

Intrinsic gas

Number of tokens in calldata