revmc_builtins::utils

Function read_words_rev

pub(crate) unsafe fn read_words_rev<'a, const N: usize>(
    sp: *mut EvmWord,
) -> &'a mut [EvmWord; N]

Splits the stack pointer into N elements by casting it to an array.

NOTE: this returns the arguments in reverse order. Use read_words! to get them in order.

The returned lifetime is valid for the entire duration of the builtin.

Safety

Caller must ensure that N matches the number of elements popped in JIT code.